backend/pyproject.toml
[project]
name = "bafangce-financial-cockpit"
version = "0.1.0"
description = "Personal financial cockpit"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "akshare==1.18.35",
    "pandas>=2.2.0",
    "numpy>=1.26.0",
    "baostock>=0.8.9",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "sqlalchemy>=2.0.0",
    "alembic>=1.14.0",
    "pydantic>=2.9.0",
    "pydantic-settings>=2.6.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "python-multipart>=0.0.12",
    "redis>=5.2.0",
    "httpx>=0.28.0",
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
]
